[Matthias/LS] Change and fix the data being shown
I now show the analytic curvature along the analytic curve, the
curvature as computed from the spline fit to the sampled analytic curve
along this spline and, as usual, the curvature as computed from the
spline fit to the pixel-digitized curve along this spline fit.

I realised that I was *wrongly!* plotting the analytical radius
distribution twice... By mistake I was never using the one computed
from the spline over the sampled curve... Now, that is fixed, and, as
expected, I have a "point-cloud" surrounding the true distribution
(for appropriate p). This is just an effect of ill-posed diff. and
smoothing the spline helps. For instance, if we try a sample spacing
of 1e-5, a p=0.(9) with a bunch of nines (9 or 10) gives good results
whereas p=1 doesn't.