[Matthias/LS] Change code,start at any polar angle
I also made some changes to the titles of the graphs: polar angle range
for original spiral, two-lined for better readability. Changed the
name of some vars to better reflect we are not necessarily using
Gaussian filters and added the ability to pinpoint the spline knots for
the pixel-precision smoothing spline.