[Matthias] Add a bunch of requested experiments
RNSE as we vary resolution and tolerance when interpolating and
smoothing, respectively. Animation of the spiral evolution w/ smoothing
tolerance. Radius as a function of the arc length when rotating both
the original image (akin to rotating the sample relative to microscope)
and the digitised image. Rounding only after and before and after
rotating, to show that, for the latter, the envelope is still formed,
just around the curve resulting from the digitised image, not the
original shape as for the former.